proxy-server/requirements.txt
flask
flask_restful
flask_cors
requests
python-dotenv